Fixed issue with cache shared across tests, moved dependency tests together
