+++ spring-cloud-zookeeper-discovery/src/main/java/org/springframework/cloud/zookeeper/discovery/dependency/DependencyRibbonAutoConfiguration.java
@@ -38,7 +38,7 @@ import org.springframework.context.annotation.Configuration;
 @AutoConfigureBefore(RibbonAutoConfiguration.class)
 @ConditionalOnRibbonZookeeper
 @Configuration
-public class ZookeeperDependencyRibbonLoadBalancerClientConfiguration {
+public class DependencyRibbonAutoConfiguration {
 
 	@Bean
 	@ConditionalOnMissingBean
