+++ spring-cloud-zookeeper-discovery/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,6 @@
 # Auto Configuration
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.cloud.zookeeper.discovery.dependency.ZookeeperDependencyRibbonLoadBalancerClientConfiguration,\
+org.springframework.cloud.zookeeper.discovery.dependency.DependencyRibbonAutoConfiguration,\
 org.springframework.cloud.zookeeper.discovery.RibbonZookeeperAutoConfiguration,\
 org.springframework.cloud.zookeeper.discovery.dependency.ZookeeperDependenciesAutoConfiguration,\
 org.springframework.cloud.zookeeper.discovery.watcher.DependencyWatcherAutoConfiguration
